+++ src/core/index.ts
@@ -1,3 +1,3 @@
-import { Virtualize } from "./virtualize";
-
+export { CellUtilities } from './cell.utilities'
 export { Virtualize } from './virtualize';
+export { ScrollUtilities } from './scroll.utilities';
