+++ src/core/virtualize.ts
@@ -1,3 +1,6 @@
+/**
+ * Virtualize core.
+ */
 export class Virtualize {
     
 }
