+++ src/interface/index.ts
@@ -4,3 +4,4 @@ export { IGridInstance } from './grid-instance';
 export { IGridRenderer } from './igrid-renderer'
 export { IGridRenderColumn } from './igrid-renderer-column';
 export { ICellConfig } from './icell-config';
+export { IGridEvents } from './grid-events';
